Added a job to detect if a PR was created from fork
Added a job to detect if a PR was created from fork, plus, added it as a dependency to the other jobs.
I have merged the workflows into 1 file because we can't specify job depenency between jobs that are running in different workflows. We want to first check the PR source, then, if it is valid - to run all the other jobs (more info here - https://github.com/orgs/community/discussions/26632)